Moved utility functions to more appropriate places.
