+++ drake/util/drakeGradientUtil.h
@@ -128,4 +128,59 @@ template<int QSubvectorSize, typename DerivedDM, typename DerivedDMSub>
 DLLEXPORT void setSubMatrixGradient(Eigen::MatrixBase<DerivedDM>& dM, const Eigen::MatrixBase<DerivedDMSub>& dM_submatrix, int row, int col, typename DerivedDM::Index M_rows,
     typename DerivedDM::Index q_start = 0, typename DerivedDM::Index q_subvector_size = QSubvectorSize);
 
+template <typename Derived>
+Eigen::Matrix<typename Derived::Scalar::Scalar, Derived::RowsAtCompileTime, Derived::ColsAtCompileTime> autoDiffToValueMatrix(const Eigen::MatrixBase<Derived>& auto_diff_matrix) {
+  Eigen::Matrix<typename Derived::Scalar::Scalar, Derived::RowsAtCompileTime, Derived::ColsAtCompileTime> ret(auto_diff_matrix.rows(), auto_diff_matrix.cols());
+  for (int i = 0; i < auto_diff_matrix.rows(); i++) {
+    for (int j = 0; j < auto_diff_matrix.cols(); ++j) {
+      ret(i, j) = auto_diff_matrix(i, j).value();
+    }
+  }
+  return ret;
+};
+
+template<typename Derived>
+typename Gradient<Eigen::Matrix<typename Derived::Scalar::Scalar, Derived::RowsAtCompileTime, Derived::ColsAtCompileTime>, Eigen::Dynamic>::type autoDiffToGradientMatrix(
+        const Eigen::MatrixBase<Derived>& auto_diff_matrix, int num_variables = Eigen::Dynamic)
+{
+  int num_variables_from_matrix = 0;
+  for (int i = 0; i < auto_diff_matrix.size(); ++i) {
+    num_variables_from_matrix = std::max(num_variables_from_matrix, static_cast<int>(auto_diff_matrix(i).derivatives().size()));
+  }
+  if (num_variables == Eigen::Dynamic) {
+    num_variables = num_variables_from_matrix;
+  }
+  else if (num_variables_from_matrix != 0 && num_variables_from_matrix != num_variables) {
+    std::stringstream buf;
+    buf << "Input matrix has derivatives w.r.t " << num_variables_from_matrix << " variables, whereas num_variables is " << num_variables << ".\n";
+    buf << "Either num_variables_from_matrix should be zero, or it should match num_variables.";
+    throw std::runtime_error(buf.str());
+  }
+
+  typename Gradient<Eigen::Matrix<typename Derived::Scalar::Scalar, Derived::RowsAtCompileTime, Derived::ColsAtCompileTime>, Eigen::Dynamic>::type gradient(auto_diff_matrix.size(), num_variables);
+  for (int row = 0; row < auto_diff_matrix.rows(); row++) {
+    for (int col = 0; col < auto_diff_matrix.cols(); col++) {
+      auto gradient_row = gradient.row(row + col * auto_diff_matrix.rows()).transpose();
+      if (auto_diff_matrix(row, col).derivatives().size() == 0) {
+        gradient_row.setZero();
+      } else {
+        gradient_row = auto_diff_matrix(row, col).derivatives();
+      }
+    }
+  }
+  return gradient;
+}
+
+template<typename DerivedGradient, typename DerivedAutoDiff>
+void gradientMatrixToAutoDiff(const Eigen::MatrixBase<DerivedGradient>& gradient, Eigen::MatrixBase<DerivedAutoDiff>& auto_diff_matrix)
+{
+  int nx = gradient.cols();
+  for (size_t row = 0; row < auto_diff_matrix.rows(); row++) {
+    for (size_t col = 0; col < auto_diff_matrix.cols(); col++) {
+      auto_diff_matrix(row, col).derivatives().resize(nx, 1);
+      auto_diff_matrix(row, col).derivatives() = gradient.row(row + col * auto_diff_matrix.rows()).transpose();
+    }
+  }
+}
+
 #endif /* DRAKEGRADIENTUTIL_H_ */
